simulation/pyproject.toml
[tool.poetry]
name = "simulation"
version = "0.1.0"
description = ""
authors = ["Bin Wang <binwang@umass.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
simpy = "^4.0.1"
pydantic = "^2.0.3"
pyyaml = "^6.0.1"
numpy = "^1.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py310"]
